+++ js_libraries/types/types/main-thread/events.d.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import { LynxEventPropsBase, BaseTouchEvent, BaseMouseEvent, BaseWheelEvent, BaseKeyEvent } from '../common/events';
+import { LynxEventPropsBase, BaseTouchEvent, BaseMouseEvent, BaseWheelEvent, BaseKeyEvent, BaseAnimationEvent, BaseTransitionEvent, LayoutChangeDetailEvent, UIAppearanceDetailEvent  } from '../common/events';
 import { Element } from './element';
 
 // worklet event
@@ -10,11 +10,18 @@ export interface TouchEvent extends BaseTouchEvent<Element> {}
 export interface MouseEvent extends BaseMouseEvent<Element> {}
 export interface WheelEvent extends BaseWheelEvent<Element> {}
 export interface KeyEvent extends BaseKeyEvent<Element> {}
+
+export interface AnimationEvent extends BaseAnimationEvent<Element> {}
+export interface TransitionEvent extends BaseTransitionEvent<Element> {}
+
+export interface LayoutChangeEvent extends LayoutChangeDetailEvent<Element> {}
+export interface UIAppearanceEvent extends UIAppearanceDetailEvent<Element> {}
+
 type LynxWorkletEventPropsImpl = {
   [K in keyof LynxEventPropsBase<Element> as Lowercase<`main-thread:${K}`>]: LynxEventPropsBase<Element>[K];
 };
 export interface LynxWorkletEventProps extends LynxWorkletEventPropsImpl {}
 
-declare module '../common/props.d.ts' {
+declare module '../common/props' {
   interface StandardProps extends LynxWorkletEventProps {}
 }