+++ js_libraries/types/vitest.config.ts
@@ -3,5 +3,6 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     name: 'lynx-types',
+    include: ['test/**/*.test-d.ts'],
   },
 });